#28133 properly display unicode chars in app windows and Shotgun menu Set codec to utf-8 so when unicode characters are used in fields that get displayed in an app UI, they are now correctly displayed as the intended characters rather than incoherent glyphs.
Cast the menu labels to unicode to preserve non-ascii text. Pymel expects a unicode object to be passed in to create the menu label but previously we were sending it as a string. When the Toolkit context contains non-ascii characters, for example if an Asset name is in Japanese, this caused garbled characters to display in the menu. By casting the context string as unicode, any non-ascii characters will be preserved in the menu.

Also fixes warning message for 2013 on Linux

Closes #6